deb: Bring-up to cm-10.2.
This change depends on flo change 296422e8, which creates a common
init.rc, to avoid duplicate changes between the two repos.

extract-files.sh doesn't work _exactly_ as advertised on an
the stock image; two files have to be copied (as the superuser) off of
/data/misc/audio into /system/etc/firmware/wcd9310/

mbhc.bin has to be renamed to wcd9310_mbhc.bin as well.

The files in /system/etc/firmware/wcd9310/* are in fact symlinks (on
the stock image) into /data/misc (see [1]).

[1] https://android.googlesource.com/device/asus/flo/+/72fa253ef0239492e68a9e18b59f0dc2209298e7/self-extractors/qcom/staging/proprietary/Android.mk